move things to adapter-components
