+++ packages/adapter-components/index.ts
@@ -21,6 +21,7 @@ export * as deployment from './src/deployment'
 export * as fetch from './src/fetch'
 export * as filters from './src/filters'
 export * as filterUtils from './src/filter_utils'
+export * as e2eUtils from './src/e2e_utils'
 export { FixElementsArgs, FixElementsHandler } from './src/fix_elements'
 export * as references from './src/references'
 export * from './src/add_alias'
