+++ packages/confluence-adapter/.vscode/launch.json
@@ -13,6 +13,19 @@
       "outFiles": [],
       "console": "integratedTerminal",
       "internalConsoleOptions": "neverOpen"
+    },
+    {
+      "type": "node",
+      "request": "launch",
+      "name": "confluence Debug Jest E2E Tests",
+      "cwd": "${workspaceFolder}",
+      "env": {
+        "RUN_E2E_TESTS": "1"
+      },
+      "args": ["${workspaceRoot}/node_modules/.bin/jest", "--runInBand", "--config", "${workspaceRoot}/jest.config.js"],
+      "outFiles": [],
+      "console": "integratedTerminal",
+      "internalConsoleOptions": "neverOpen"
     }
   ]
 }
